[#31] 주문 정보 가져오기 (#49)
* feature: 주문 정보 가져오기

* refactor: 변수명 변경 및 캐싱

+++ src/main/java/com/coupang/marketplace/product/controller/dto/SimpleProduct.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigInteger;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -20,7 +19,7 @@ public class SimpleProduct {
     private String name;
 
     @NotNull
-    private BigInteger price;
+    private long price;
 
     @NotNull
     private String mainImg;
@@ -29,7 +28,7 @@ public class SimpleProduct {
     private float score;
 
     @Builder
-    public SimpleProduct(long id, String name, BigInteger price, String mainImg, float score) {
+    public SimpleProduct(long id, String name, long price, String mainImg, float score) {
         this.id = id;
         this.name = name;
         this.price = price;
